Fixed bug in RecoverySegmentBuilder that created multiple output copies of each safe-version entry, causing logs to double or triple in size with each recovery. Eventually, there wasn't enough space in the output segment to hold all of the safe-version entries from the input segment.

+++ src/BackupMasterRecovery.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2009-2012 Stanford University
+/* Copyright (c) 2009-2013 Stanford University
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -83,7 +83,7 @@ class BackupMasterRecovery : public Task {
     void setPartitionsAndSchedule(ProtoBuf::Tablets partitions);
     Status getRecoverySegment(uint64_t recoveryId,
                               uint64_t segmentId,
-                              uint64_t partitionId,
+                              int partitionId,
                               Buffer* buffer,
                               Segment::Certificate* certificate);
     void free();
@@ -136,7 +136,7 @@ class BackupMasterRecovery : public Task {
      * in #partitions may have the same partition id. One more than the
      * largest id found among the partitions.
      */
-    size_t numPartitions;
+    int numPartitions;
 
     /**
      * Keeps all state for each replica that is part of this master recovery.
